shopping cart dev

+++ src/main/java/dev/a2/onlinestore/repos/UsersRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import dev.a2.onlinestore.model.User;
 
 @Repository
-public interface UsersRepository extends JpaRepository<User, Integer> {
+public interface UsersRepository extends JpaRepository<User, Long> {
+	
     User findByEmail(String email);
 }
